Reset the type variable back the KVM_HC_SEND_IPI after the first vmcall
The first vmcall overwrites the type variable with it's own return
value. The second vmcall will now be done with a type that is not
necessarily KVM_HC_SEND_IPI, but whatever hypercall the return value
of the first vmcall maps to.

In my case, the first vmcall returned 4, which corresponds to the type
KVM_HC_PPC_MAP_MAGIC_PAGE, i.e. a completely different vmcall.

Of course I don't see a cache hit when I do the wrong vmcall...